Now, execute function can also be used to just get the response, without displaying it. Helpful for making chained requests. Uses a polyfill though, A global class "ResponseHandler" to save the response.
The execute() method from Core/Applications class now takes 3rd argument for ResponseType, default is 'response' which uses the polyfill method above. Although when we call the script from URL i.e. the parseURL method, it calls execute() function with the responseType as 'json', which will echo the result.

The Core/Controller has a new variable that stores this responseType and each controller has to call the view() method with this parameter, so that it is generalised in all controller calls.

That's all! Was confusing to implement, but is still very easy to use!